Tests/LibWeb/Text/expected/selection-rect-with-letter-spacing.txt

firstHalf + secondHalf: 391.6875
fullWidth: 391.6875
